requirements.txt
flask==2.3.3
python-dotenv==1.1.1
flask-cors==6.0.0
requests==2.32.4
werkzeug==3.0.6
cryptography==45.0.6
psycopg2-binary==2.9.10